Pin Animations
There are now animations for selecting, placing, and removing pins. The animation is done through PinAnimator.gd and tweens the position and rotation of the pins using Curves. If an animation must play whilst one is already playing, then the currently playing animation will be interrupted by the new one. Interrupting animations is seamless as the new animation will start from the position of the old one.