Refactored PinAnimator.gd, and Main.gd. Added a cover that hides the target pins
Refactored Main.gd to encapsulate just code for the board. Created a new helper class BoardLogic thta decouples the logic for the board from the visuals. Pin is now dependent on PinAnimator (instead of PinAnimator being dependent on Pin). This is better because classes that use Pin do not need to know how to use the PinAnimator. Created a cover from the target pins that provides methods for opening and closing the cover such that the target pins are visible.